remove op._str(), group base operators
